Move comparer to new class
